Changed unsigned int for unsigned long and double for long double

+++ lib/random/rand.hpp
@@ -7,37 +7,37 @@ namespace pix::random
      * @brief Set seed
      * @param seed Initial value
     */
-    void set_seed(unsigned int);
+    void set_seed(unsigned long);
 
     /**
      * @brief Set multiplier
      * @param mult Multiplier
     */
-    void set_mult(unsigned int);
+    void set_mult(unsigned long);
 
     /**
      * @brief Set increment
      * @param inc Increment
     */
-    void set_inc(unsigned int);
+    void set_inc(unsigned long);
 
     /**
      * @brief Set modulus
      * @param mod Modulus
     */
-    void set_mod(unsigned int);
+    void set_mod(unsigned long);
 
     /**
      * @brief Linear congruential generator
      * @return Random number between 0 and INT32_MAX
     */
-    int rand(void);
+    unsigned long rand(void);
 
     /**
      * @brief Line congruential generator
      * @return Random number between 0 and 1
     */
-    double drand(void);
+    long double drand(void);
 }
 
 // Implementation file
